notification_service/requirements.txt
fastapi
uvicorn
psycopg2
cassandra-driver
pydantic
sqlalchemy
requests